refactor: Extract default cell from default item
This will make it easier to insert a slot when the cells will differ (when we use this for social)
