+++ src/components/comparison/ComparisonRow.vue
@@ -5,9 +5,11 @@
           <div class="definition">{{ subtitle }}</div>
       </td>
       <td v-for="(study, index) in studies" :key="`value_added__${study.id}`">
-          <div :class="getAddedValueClass(values[index])">
-              {{ values[index] ? format(values[index]) : '-' }}
-          </div>
+          <ComparisonDefaultCell
+            :value="values[index]"
+            :valueType="valueType"
+            :reverseColors="reverseColors"
+          />
       </td>
       <td></td>
   </tr>
@@ -15,7 +17,7 @@
 
 <script setup>
 import { computed } from 'vue';
-import { formatNumber, formatPercent } from '@utils/format.js'
+import ComparisonDefaultCell from './ComparisonDefaultCell.vue';
 
 const props = defineProps({
     studies: Array,
@@ -30,37 +32,4 @@ const props = defineProps({
 })
 
 const values = computed(() => props.studies.map(study => props.getValue(study)))
-
-function getAddedValueClass(value) {
-  return getPositiveOrNegativeClass(props.reverseColors ? -value : value);
-}
-function getPositiveOrNegativeClass(value) {
-  if (!value) {
-      return "gray"
-  }
-  if (value < 0) {
-      return "negative"
-  }
-  return "positive"
-}
-
-function format(value) {
-  switch (props.valueType) {
-    case "percent":
-      return formatPercent(value);
-    case "number":
-      return formatNumber(value);
-    default:
-      return "";
-  }
-}
 </script>
-
-<style scoped lang="scss">
-    .negative {
-        background-color: #ffac9e;
-    }
-    .positive {
-        background-color: #94d99d;
-    }
-</style>
